major refactor: migrate to neon postgress, reorganize lib directory
