+++ drizzle.config.ts
@@ -4,11 +4,10 @@ import { config } from '@/lib/config'
 
 export default {
   schema: 'src/db/schema.ts',
-  out: 'src/db/migrations',
-  dialect: 'turso',
+  out: './migrations',
+  dialect: 'postgresql',
   dbCredentials: {
-    url: config.db.TURSO_CONNECTION_URL,
-    authToken: config.db.TURSO_AUTH_TOKEN,
+    url: config.db.URL,
   },
   verbose: true,
   strict: true,
