--- src/features/events/types.ts
@@ -1,5 +1,0 @@
-export type ConnectionStatus = 'pending' | 'success' | 'failed'
-
-export type ConnectionMessage = {
-  status: ConnectionStatus
-}
