+++ src/hooks/use-payment-channel.ts
@@ -1,7 +1,6 @@
 import { useEffect, useState } from 'react'
 
-import type { ConnectionMessage, ConnectionStatus } from '@/features/events/types'
-import type { PaymentUpdate } from '@/lib/payments/publish-update'
+import type { ConnectionMessage, ConnectionStatus, PaymentUpdate } from '@/lib/types'
 
 export default function usePaymentChannel(paymentId: string): [update: PaymentUpdate | null, concectionStatus: ConnectionStatus, error: string | null] {
   const [update, setUpdate] = useState<PaymentUpdate | null>(null)
