+++ src/lib/constants.ts
@@ -11,3 +11,5 @@ export const GENDERS = [
   'male',
   'female',
 ] as const
+
+export const PAYMENT_STATUSES = ['pending', 'success', 'failed'] as const
