+++ src/lib/payments/publish-update.ts
@@ -1,15 +1,6 @@
-import type { PaymentStatus } from '../mpesa/types'
-import type { PaymentUpdateChannel } from './types'
+import type { PaymentUpdate, PaymentUpdateChannel } from '../types'
 
 import { redisPublisher } from '../upstash-redis'
-
-export type PaymentUpdate = {
-  status: PaymentStatus
-  amount?: number
-  mpesaReceiptNumber?: string
-  transactionDate?: string
-  errorMessage?: string
-}
 
 export async function publishPaymentUpdate(paymentId: string, update: PaymentUpdate) {
   const channel: PaymentUpdateChannel = `paymentId:${paymentId}`
