--- src/lib/payments/types.ts
@@ -1,1 +1,0 @@
-export type PaymentUpdateChannel = `paymentId:${string}`
