+++ src/lib/validation-schemas.ts
@@ -39,3 +39,22 @@ export const insertParticipantSchema = createInsertSchema(participants).omit({ i
   category: z.enum(CATEGORIES),
   gender: z.enum(GENDERS),
 })
+
+export const callbackSchema = z.object({
+  Body: z.object({
+    stkCallback: z.object({
+      MerchantRequestID: z.string(),
+      CheckoutRequestID: z.string(),
+      ResultCode: z.number(),
+      ResultDesc: z.string(),
+      CallbackMetadata: z
+        .object({
+          Item: z.array(z.object({
+            Name: z.string(),
+            Value: z.union([z.string(), z.number()]).optional(),
+          })),
+        })
+        .optional(),
+    }),
+  }),
+})
